cast to avoid bitloss, hmm
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v1_7/bzflag@582 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/bzfs.cxx
@@ -383,7 +383,7 @@ private:
       b[3] = atoi(ip);
 
 	mask.s_addr= ((unsigned int)b[0] << 24) |
-        ((unsigned int)b[1] << 16) | ((unsigned int)b[2] << 8) | b[3];
+        ((unsigned int)b[1] << 16) | ((unsigned int)b[2] << 8) | (unsigned int)b[3];
     return true;
   }
